+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+## [0.17.0] - 2019-06-16
+
 ### Changed
 
 - The DM options menu is now a proper dialog.